# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/borrows": {
            "get": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Get a list of all borrows with users and books",
                "operationId": "c3826462f8f70e7c3baa103e1159907f",
                "responses": {
                    "200": {
                        "description": "List of borrows with users and books"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Create a new borrow",
                "operationId": "be2d0b69211f03afc4ff7c5e8adbc628",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "borrowed_at"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "borrowed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrow record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrowed with success!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book could not be borrowed!"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/borrows/{id}": {
            "get": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Get a specific borrow by ID",
                "operationId": "f9f0f076e437a8b56d92f0a3d23641bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the borrow record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow record details"
                    },
                    "404": {
                        "description": "Borrow record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow record not found!"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Borrow record with the specified ID does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Update a specific borrow",
                "operationId": "2e69c4d52774fad19b54f9d1c7de1969",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the borrow",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "borrowed_at",
                                    "returned_at"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "borrowed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-26"
                                    },
                                    "returned_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrow record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrow updated successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrow could not be updated!"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Delete a specific borrow",
                "operationId": "bfc058ded486d474b980393321e1efbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the borrow",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow record deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrow deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Borrow record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow record not found!"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Borrow record with the specified ID does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrow could not be deleted!"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}